api/backend/requirements.txt
anyio==3.5.0
asgiref==3.5.0
boto3==1.20.45
botocore==1.23.45
click==8.0.3
et-xmlfile==1.1.0
fastapi==0.73.0
flake8==4.0.1
fpdf==1.7.2
greenlet==1.1.2
h11==0.13.0
idna==3.3
jmespath==0.10.0
mccabe==0.6.1
numpy==1.22.1
openpyxl==3.0.9
pandas==1.4.0
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pydantic==1.9.0
pyflakes==2.4.0
python-dateutil==2.8.2
python-dotenv==0.19.2
python-multipart==0.0.5
pytz==2021.3
s3transfer==0.5.0
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.31
starlette==0.17.1
typing-extensions==4.0.1
Unidecode==1.3.2
urllib3==1.26.8
uvicorn==0.17.1
